fixup! feat(forms): add DI option for classes on `Field` directive
Address review feedback
