+++ packages/forms/signals/src/api/di.ts
@@ -21,22 +21,6 @@ export interface SignalFormsConfig {
 }
 
 /**
- * A value that can be used for {@link SignalFormsConfig.classes} to automatically add
- * the `ng-*` status classes from reactive forms.
- *
- * @experimental 21.0.1
- */
-export const NG_STATUS_CLASSES: SignalFormsConfig['classes'] = {
-  'ng-touched': (state) => state.touched(),
-  'ng-untouched': (state) => !state.touched(),
-  'ng-dirty': (state) => state.dirty(),
-  'ng-pristine': (state) => !state.dirty(),
-  'ng-valid': (state) => state.valid(),
-  'ng-invalid': (state) => state.invalid(),
-  'ng-pending': (state) => state.pending(),
-};
-
-/**
  * Provides configuration options for signal forms.
  *
  * @experimental 21.0.1
